fix for ../fuzz/rar5out/slave/hangs/id:000002,src:001090,op:havoc,rep:4 uninit-stack
size is uninited in some cases, e.g. do_unpack() hits uncompress_file() and then
we exit. I guess this is unusual and normally we hit things in a different order.

Anyway this leaves *size with random stack data in bsdtar. Don't do that,
initalise it early.

We could set it to a canary value and detect the issue, but I don't understand
the issue and this also leaves the ability to redo a size 0 read and actually
get some data, I think.